# api/openapi.yaml
...
components:
  schemas:
    ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    ReconcileStats:
      properties:
        conversation_joined:
          description: |-
            ConversationJoined / ConversationTotal measure how many
            conversation-spine nodes' content appears in a transcript — the
            Go-native version of the prototype's join-rate oracle.
          type: integer
        conversation_total:
          type: integer
        forked_chains:
          type: integer
        main_chains_joined:
          type: integer
        subagent_forks:
          type: integer
        transcript_files:
          type: integer
      type: object
    RederiveReport:
      properties:
        attached_verdicts:
          type: integer
        call_kinds:
          additionalProperties:
            type: integer
          type: object
        judged_actions:
          description: |-
            Verdict attach: judged actions grouped across stages, and how
            many attached one-to-one to a captured tool_use.
          type: integer
        node_kinds:
          additionalProperties:
            type: integer
          type: object
        nodes:
          type: integer
        parse_failures:
          items:
            type: string
          type: array
        parsed_turns:
          type: integer
        plans_attached:
          description: |-
            PlansAttached counts plan-name-gen calls linked to the
            ExitPlanMode tool_use that accepted the plan.
          type: integer
        raw_only_turns:
          type: integer
        raw_turns:
          type: integer
        reconcile:
          allOf:
          - $ref: '#/components/schemas/ReconcileStats'
          description: |-
            Reconcile reports the transcript↔wire fusion when transcript
            rows are present in the raw layer.
        unattached_actions:
          description: |-
            UnattachedActions samples judged actions that found no matching
            tool_use (capped) — expected for non-tool events like subagent
            handbacks; anything else is matcher signal worth reading.
          items:
            type: string
          type: array
        web_summary_attached:
          description: |-
            WebSummaryAttached counts web-summary calls linked back to their
            WebFetch/WebSearch tool_use.
          type: integer
      type: object
...
    SpanSearchOutput:
      properties:
        count:
          type: integer
        query:
          type: string
        results:
          items:
            $ref: '#/components/schemas/SpanSearchResult'
          type: array
      type: object
    SpanSearchResult:
      properties:
        model:
          type: string
        score:
          type: number
        session_id:
          type: string
        snippet:
          description: Snippet previews the matched span's delta-only text.
          type: string
        span_id:
          type: string
        started_at:
          type: string
        trace_id:
          type: string
        user_prompt:
          description: |-
            UserPrompt is the prompt of the turn (trace) the span belongs to.
            Served explicitly (not omitempty) so a synthetic turn's empty prompt
            reaches consumers as "" rather than a dropped key — see TraceItem.
          type: string
      type: object
...
    deriveRunResponse:
      properties:
        orgs:
          additionalProperties:
            $ref: '#/components/schemas/RederiveReport'
          type: object
      type: object
...
    github_com_papercomputeco_tapes_pkg_seed.Result:
      properties:
        raw_turns:
          description: RawTurns is the total number of corpus rows replayed.
          type: integer
        raw_turns_deduped:
          type: integer
        raw_turns_inserted:
          description: |-
            RawTurnsInserted counts rows that landed as new raw turns;
            RawTurnsDeduped counts replays the raw layer's dedup absorbed
            (a re-seed reports everything deduped).
          type: integer
        sessions:
          description: Sessions is the number of demo sessions the corpora replay
            into.
          type: integer
      type: object
...
    seedDemoRequest:
      properties:
        overwrite:
          type: boolean
      type: object
...
info:
  contact: {}
  description: |-
    HTTP API for inspecting, querying, and searching stored Tapes sessions.

    The REST surface exposes health checks, session listing and retrieval, derived session summaries, aggregate stats, semantic search, skill authoring and publishing, operator maintenance endpoints, and a streamable MCP endpoint.

    Not covered here: the ingest write surface, which is a separate server with its own contract (see ingest/openapi.yaml).
  title: Tapes API
  version: "1.0"
openapi: 3.0.3
paths:
...
  /v1/admin/derive/run:
    post:
      description: |-
        Rebuilds traces, spans, links, and session rollups for every org from the immutable raw-turn store. Idempotent: re-running reproduces the same projection and prunes rows the current derive no longer emits.

        This is how a projection or classifier change reaches already-captured data — it re-derives rather than re-captures. Cost scales with the raw layer, so it is an operator lever, not a request-path call.
      operationId: runDerive
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deriveRunResponse'
          description: Per-org derive reports
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Derive failed
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Driver does not host the raw-turn layer
      summary: Re-derive the span projection (operator)
      tags:
      - admin
  /v1/admin/seed/demo:
    post:
      description: 'Replays the bundled demo capture corpora through the ingest write
        path into the caller''s org, then derives the seeded sessions. Idempotent:
        raw-turn dedup makes repeat seeds no-ops.'
      operationId: seedDemo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/seedDemoRequest'
        description: Seed options (overwrite is no longer supported)
        x-originalParamName: request
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_papercomputeco_tapes_pkg_seed.Result'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid payload or unsupported option
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Seeding failed
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Driver does not host the raw-turn layer
      summary: Seed demo sessions (operator)
      tags:
      - admin
...
  /v1/search/spans:
    get:
      description: Embeds the query text and runs vector similarity over the embedded
        span projection (main llm spans, delta-only content). Each hit carries span,
        trace, and turn context.
      operationId: searchSpans
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        schema:
          type: string
      - description: Maximum number of results to return
        in: query
        name: top_k
        schema:
          default: 5
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpanSearchOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing or invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Search execution failed
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Span search is not configured or not yet initialized
      summary: Semantic search over span embeddings
      tags:
      - search
...